Made the datamodel inherit all of the database data so we can do some iterative route solving
put false_delivery handling in the API

Can't do much else though in terms of processing the data
the problem with solving continuously is that the routecalculator
only knows one instance of data (like one instance of plane location)
so it can't process the output on its own

For example, we don't know when to cross off the finished deliveries
because we don't know the full route of the plane, etc.

We need to do this stuff in the controller or something that knows
the previous plane path, etc.